refactor code: created a `set_post_info` func, which sets variables that other funcs can use to print post related HTML (post listings, indexes, post page, etc)
Also added a new function (not yet used) called `print_post_preview` .. Give it the path to a post `.md` file, and it will print a preview of that post (for use in post listings. The preview shows title, data, time to read, and truncated description.

This new `print_post_preview` function takes an MD file as $1, and uses the new `set_post_info` func to set the vars it uses in the HTML it outputs such as $post_title, $post_descr and so on.

TLDR: This commit cleans up the code a bit and makes it easier to grab/set post info before creating post related HTML.